renamed sample files to follow new scheme
