+++ src/frontend-samples/app-ui-samples/toolbar-button-provider-sample/ToolbarButtonApp.tsx
@@ -3,9 +3,7 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import React, { FunctionComponent, useState } from "react";
-import { ToolbarButtonProvider } from "./ToolbarButtonUi";
-import "common/samples-common.scss";
-import "common/AppUi/app-ui.scss";
+import { ToolbarButtonProvider } from "./ToolbarButtonProvider";
 import { AuthorizationClient, default3DAppUi, useSampleWidget, ViewSetup } from "@itwinjs-sandbox";
 import { IModelConnection } from "@bentley/imodeljs-frontend";
 import { Viewer } from "@bentley/itwin-viewer-react";
@@ -13,7 +11,7 @@ import { IModelViewportControlOptions } from "@bentley/ui-framework";
 
 const uiProviders = [new ToolbarButtonProvider()];
 
-const ToolbarButtonSample: FunctionComponent = () => {
+const ToolbarButtonApp: FunctionComponent = () => {
   const sampleIModelInfo = useSampleWidget("Press the Lightbulb button tool at the top of the screen.");
   const [viewportOptions, setViewportOptions] = useState<IModelViewportControlOptions>();
 
@@ -41,4 +39,4 @@ const ToolbarButtonSample: FunctionComponent = () => {
   );
 };
 
-export default ToolbarButtonSample;
+export default ToolbarButtonApp;
