+++ src/frontend-samples/image-export/ImageExportApp.tsx
@@ -2,23 +2,44 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+import React, { FunctionComponent, useState } from "react";
+import { AuthorizationClient, default3DSandboxUi, SampleIModels, useSampleWidget, ViewSetup } from "@itwinjs-sandbox";
+import { Viewer } from "@bentley/itwin-viewer-react";
+import { IModelConnection } from "@bentley/imodeljs-frontend";
+import { IModelViewportControlOptions } from "@bentley/ui-framework";
+import { ImageExportWidgetProvider } from "./ImageExportWidget";
 
-import "common/samples-common.scss";
-import { IModelApp } from "@bentley/imodeljs-frontend";
+const uiProviders = [new ImageExportWidgetProvider()];
 
-export default class ImageExportApp {
+const ImageExportApp: FunctionComponent = () => {
+  const sampleIModelInfo = useSampleWidget("Use Image Export Controls Widget to export current viewport as an image.", [SampleIModels.House, SampleIModels.MetroStation]);
+  const [viewportOptions, setViewportOptions] = useState<IModelViewportControlOptions>();
 
-  // Capture image of currently active viewport and trigger browser's download action.
-  public static exportImage() {
-    const viewPort = IModelApp.viewManager.getFirstOpenView();
-    if (viewPort !== undefined) {
-      const canvas = viewPort.readImageToCanvas();
-      const imageUrl = canvas.toDataURL("image/png");
-      const link = document.createElement("a");
-      link.setAttribute("download", "viewport.png");
-      link.setAttribute("href", imageUrl);
-      link.click();
-    }
-  }
+  const _oniModelReady = async (iModelConnection: IModelConnection) => {
+    const viewState = await ViewSetup.getDefaultView(iModelConnection);
+    setViewportOptions({ viewState });
+  };
 
-}
+  /** The sample's render method */
+  return (
+    <>
+      { /** Viewport to display the iModel */}
+      {sampleIModelInfo?.iModelName && sampleIModelInfo?.contextId && sampleIModelInfo?.iModelId &&
+        <Viewer
+          contextId={sampleIModelInfo.contextId}
+          iModelId={sampleIModelInfo.iModelId}
+          authConfig={{ oidcClient: AuthorizationClient.oidcClient }}
+          viewportOptions={viewportOptions}
+          onIModelConnected={_oniModelReady}
+          defaultUiConfig={default3DSandboxUi}
+          theme="dark"
+          uiProviders={uiProviders}
+        />
+      }
+    </>
+  );
+
+};
+
+export default ImageExportApp;
+
