+++ src/frontend-samples/view-attributes-sample/ViewAttributes.scss
@@ -6,3 +6,11 @@
   color: white;
   padding: 8px;
 }
+
+.sample-options-2col {
+  display: grid;
+  row-gap: 8px;
+  column-gap: 8px;
+  grid-template-columns: 1fr 2fr;
+  align-items: center;
+}
