+++ src/frontend-samples/view-attributes-sample/ViewAttributesApp.tsx
@@ -2,185 +2,48 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+import { AuthorizationClient, default3DSandboxUi, SampleIModels, useSampleWidget, ViewSetup } from "@itwinjs-sandbox";
+import React, { FunctionComponent, useState } from "react";
+import { Viewer } from "@bentley/itwin-viewer-react";
+import { ViewAttributesWidgetProvider } from "./ViewAttributesWidget";
+import { IModelApp, IModelConnection, ScreenViewport } from "@bentley/imodeljs-frontend";
+import { ViewAttributesApi } from "./ViewAttributesApi";
+import { IModelViewportControlOptions } from "@bentley/ui-framework";
 
-import "common/samples-common.scss";
-import { Environment, Viewport, ViewState3d } from "@bentley/imodeljs-frontend";
-import { RenderMode } from "@bentley/imodeljs-common";
+const uiProviders = [new ViewAttributesWidgetProvider()];
 
-// cSpell:ignore imodels
+const ViewAttributesApp: FunctionComponent = () => {
+  const sampleIModelInfo = useSampleWidget("Use the controls below to change the view attributes.", [SampleIModels.House, SampleIModels.MetroStation]);
+  const [viewportOptions, setViewportOptions] = useState<IModelViewportControlOptions>();
 
-export enum ViewFlag {
-  ACS, BackgroundMap, Grid, HiddenEdges, Monochrome, VisibleEdges, Shadows,
-}
+  const _oniModelReady = async (iModelConnection: IModelConnection) => {
+    IModelApp.viewManager.onViewOpen.addOnce(async (_vp: ScreenViewport) => {
+      ViewAttributesApi.setAttrValues(_vp, ViewAttributesApi.settings);
+    });
 
-export interface AttrValues {
-  renderMode: RenderMode;
-  acs: boolean;
-  backgroundMap: boolean;
-  backgroundTransparency: number | false;
-  cameraOn: boolean;
-  grid: boolean;
-  hiddenEdges: boolean;
-  monochrome: boolean;
-  shadows: boolean;
-  skybox: boolean;
-  visibleEdges: boolean;
-}
-
-/** This class implements the interaction between the sample and the iModel.js API.  No user interface. */
-export default class ViewAttributesApp {
-
-  public static settings: AttrValues = {
-    renderMode: RenderMode.SmoothShade,
-    acs: false,
-    backgroundMap: true,
-    backgroundTransparency: 0.01,
-    cameraOn: true,
-    grid: false,
-    hiddenEdges: false,
-    monochrome: false,
-    shadows: false,
-    skybox: true,
-    visibleEdges: false,
+    const viewState = await ViewSetup.getDefaultView(iModelConnection);
+    setViewportOptions({ viewState });
   };
 
-  public static getAttrValues(vp: Viewport): AttrValues {
-    return {
-      renderMode: ViewAttributesApp.getRenderModel(vp),
-      acs: ViewAttributesApp.getViewFlag(vp, ViewFlag.ACS),
-      backgroundMap: ViewAttributesApp.getViewFlag(vp, ViewFlag.BackgroundMap),
-      backgroundTransparency: ViewAttributesApp.getBackgroundTransparency(vp),
-      cameraOn: ViewAttributesApp.isCameraOn(vp),
-      grid: ViewAttributesApp.getViewFlag(vp, ViewFlag.Grid),
-      hiddenEdges: ViewAttributesApp.getViewFlag(vp, ViewFlag.HiddenEdges),
-      monochrome: ViewAttributesApp.getViewFlag(vp, ViewFlag.Monochrome),
-      shadows: ViewAttributesApp.getViewFlag(vp, ViewFlag.Shadows),
-      skybox: ViewAttributesApp.isSkyboxOn(vp),
-      visibleEdges: ViewAttributesApp.getViewFlag(vp, ViewFlag.VisibleEdges),
-    };
-  }
+  /** The sample's render method */
+  return (
+    <>
+      { /** Viewport to display the iModel */}
+      {sampleIModelInfo?.iModelName && sampleIModelInfo?.contextId && sampleIModelInfo?.iModelId &&
+        <Viewer
+          contextId={sampleIModelInfo.contextId}
+          iModelId={sampleIModelInfo.iModelId}
+          authConfig={{ oidcClient: AuthorizationClient.oidcClient }}
+          viewportOptions={viewportOptions}
+          onIModelConnected={_oniModelReady}
+          defaultUiConfig={default3DSandboxUi}
+          theme="dark"
+          uiProviders={uiProviders}
+        />
+      }
+    </>
+  );
 
-  public static setAttrValues(vp: Viewport, attrValues: AttrValues) {
-    const currAttrValues = this.getAttrValues(vp);
+};
 
-    if (currAttrValues.renderMode !== attrValues.renderMode)
-      ViewAttributesApp.setRenderMode(vp, attrValues.renderMode);
-    if (attrValues.backgroundTransparency && currAttrValues.backgroundTransparency !== attrValues.backgroundTransparency)
-      ViewAttributesApp.setBackgroundTransparency(vp, attrValues.backgroundTransparency);
-    if (currAttrValues.cameraOn !== attrValues.cameraOn)
-      ViewAttributesApp.setCameraOnOff(vp, attrValues.cameraOn);
-    if (currAttrValues.skybox !== attrValues.skybox)
-      ViewAttributesApp.setSkyboxOnOff(vp, attrValues.skybox);
-
-    // Update viewflags
-    vp.viewFlags.acsTriad = attrValues.acs;
-    vp.viewFlags.backgroundMap = attrValues.backgroundMap;
-    vp.viewFlags.grid = attrValues.grid;
-    vp.viewFlags.hiddenEdges = attrValues.hiddenEdges;
-    vp.viewFlags.monochrome = attrValues.monochrome;
-    vp.viewFlags.shadows = attrValues.shadows;
-    vp.viewFlags.visibleEdges = attrValues.visibleEdges;
-
-    vp.synchWithView();
-  }
-
-  // Query flag values using the Viewport API.
-  public static getViewFlag(vp: Viewport, flag: ViewFlag): boolean {
-    switch (flag) {
-      case ViewFlag.ACS: return vp.viewFlags.acsTriad;
-      case ViewFlag.BackgroundMap: return vp.viewFlags.backgroundMap;
-      case ViewFlag.Grid: return vp.viewFlags.grid;
-      case ViewFlag.HiddenEdges: return vp.viewFlags.hiddenEdges;
-      case ViewFlag.Monochrome: return vp.viewFlags.monochrome;
-      case ViewFlag.Shadows: return vp.viewFlags.shadows;
-      case ViewFlag.VisibleEdges: return vp.viewFlags.visibleEdges;
-    }
-  }
-
-  // Modify flag values using the Viewport API.
-  public static setViewFlag(vp: Viewport, flag: ViewFlag, on: boolean) {
-    switch (flag) {
-      case ViewFlag.ACS:
-        vp.viewFlags.acsTriad = on;
-        break;
-      case ViewFlag.BackgroundMap:
-        vp.viewFlags.backgroundMap = on;
-        break;
-      case ViewFlag.Grid:
-        vp.viewFlags.grid = on;
-        break;
-      case ViewFlag.HiddenEdges:
-        vp.viewFlags.hiddenEdges = on;
-        break;
-      case ViewFlag.Monochrome:
-        vp.viewFlags.monochrome = on;
-        break;
-      case ViewFlag.Shadows:
-        vp.viewFlags.shadows = on;
-        break;
-      case ViewFlag.VisibleEdges:
-        vp.viewFlags.visibleEdges = on;
-        break;
-    }
-    vp.synchWithView();
-  }
-
-  // Query camera setting using the Viewport API.
-  public static isCameraOn(vp: Viewport) {
-    return vp.isCameraOn;
-  }
-
-  // Query map background transparency using the Viewport API
-  public static getBackgroundTransparency(vp: Viewport) {
-    return vp.backgroundMapSettings.transparency;
-  }
-
-  // Modify map background transparency using the Viewport API
-  public static setBackgroundTransparency(vp: Viewport, transparency: number) {
-    const style = vp.backgroundMapSettings.clone({ transparency });
-    vp.displayStyle.backgroundMapSettings = style;
-    vp.synchWithView();
-  }
-
-  // Modify camera setting using the Viewport API.
-  public static setCameraOnOff(vp: Viewport, on: boolean) {
-    if (on)
-      vp.turnCameraOn();
-    else
-      (vp.view as ViewState3d).turnCameraOff();
-
-    vp.synchWithView();
-  }
-
-  // Query skybox setting using the Viewport API.
-  public static isSkyboxOn(vp: Viewport) {
-    if (vp.view.is3d()) {
-      const displayStyle = vp.view.getDisplayStyle3d();
-      return displayStyle.environment.sky.display;
-    }
-
-    return false;
-  }
-
-  // Modify skybox setting using the Viewport API.
-  public static setSkyboxOnOff(vp: Viewport, on: boolean) {
-    if (vp.view.is3d()) {
-      const style = vp.view.getDisplayStyle3d();
-      style.environment = new Environment({ sky: { display: on } });
-      vp.invalidateRenderPlan();
-    }
-  }
-
-  // Query render model setting using the Viewport API.
-  public static getRenderModel(vp: Viewport): RenderMode {
-    return vp.viewFlags.renderMode;
-  }
-
-  // Modify render mode setting using the Viewport API.
-  public static setRenderMode(vp: Viewport, mode: RenderMode) {
-    const viewFlags = vp.viewFlags.clone();
-    viewFlags.renderMode = mode;
-    vp.viewFlags = viewFlags;
-  }
-
-}
+export default ViewAttributesApp;
