--- src/frontend-samples/view-attributes-sample/ViewAttributesUI.tsx
@@ -1,49 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-import { AuthorizationClient, default3DSandboxUi, SampleIModels, useSampleWidget, ViewSetup } from "@itwinjs-sandbox";
-import React, { FunctionComponent, useState } from "react";
-import { Viewer } from "@bentley/itwin-viewer-react";
-import { ViewAttributesWidgetProvider } from "./ViewAttributesWidget";
-import { IModelApp, IModelConnection, ScreenViewport } from "@bentley/imodeljs-frontend";
-import ViewAttributesApp from "./ViewAttributesApp";
-import { IModelViewportControlOptions } from "@bentley/ui-framework";
-
-const uiProviders = [new ViewAttributesWidgetProvider()];
-
-const ViewAttributesUI: FunctionComponent = () => {
-  const sampleIModelInfo = useSampleWidget("Use the controls below to change the view attributes.", [SampleIModels.House, SampleIModels.MetroStation]);
-  const [viewportOptions, setViewportOptions] = useState<IModelViewportControlOptions>();
-
-  const _oniModelReady = async (iModelConnection: IModelConnection) => {
-    IModelApp.viewManager.onViewOpen.addOnce(async (_vp: ScreenViewport) => {
-      ViewAttributesApp.setAttrValues(_vp, ViewAttributesApp.settings);
-    });
-
-    const viewState = await ViewSetup.getDefaultView(iModelConnection);
-    setViewportOptions({ viewState });
-  };
-
-  /** The sample's render method */
-  return (
-    <>
-      { /** Viewport to display the iModel */}
-      {sampleIModelInfo?.iModelName && sampleIModelInfo?.contextId && sampleIModelInfo?.iModelId &&
-        <Viewer
-          contextId={sampleIModelInfo.contextId}
-          iModelId={sampleIModelInfo.iModelId}
-          authConfig={{ oidcClient: AuthorizationClient.oidcClient }}
-          viewportOptions={viewportOptions}
-          onIModelConnected={_oniModelReady}
-          defaultUiConfig={default3DSandboxUi}
-          theme="dark"
-          uiProviders={uiProviders}
-        />
-      }
-    </>
-  );
-
-};
-
-export default ViewAttributesUI;
