+++ src/frontend-samples/viewer-only-2d-sample/ViewerOnly2d.scss
