+++ src/frontend-samples/viewport-only-sample/ViewportOnlyApp.tsx
@@ -2,17 +2,14 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import { AuthorizationClient, default3DSandboxUi, SampleIModels, useSampleWidget, ViewSetup } from "@itwinjs-sandbox";
 import React, { FunctionComponent, useState } from "react";
+import { AuthorizationClient, default3DSandboxUi, useSampleWidget, ViewSetup } from "@itwinjs-sandbox";
 import { Viewer } from "@bentley/itwin-viewer-react";
 import { IModelConnection } from "@bentley/imodeljs-frontend";
 import { IModelViewportControlOptions } from "@bentley/ui-framework";
-import { ImageExportWidgetProvider } from "./ImageExportWidget";
 
-const uiProviders = [new ImageExportWidgetProvider()];
-
-const ImageExportUI: FunctionComponent = () => {
-  const sampleIModelInfo = useSampleWidget("Use the controls below to change the view attributes.", [SampleIModels.House, SampleIModels.MetroStation]);
+const ViewportOnlyApp: FunctionComponent = () => {
+  const sampleIModelInfo = useSampleWidget("Use the toolbar at the top-right to navigate the model.");
   const [viewportOptions, setViewportOptions] = useState<IModelViewportControlOptions>();
 
   const _oniModelReady = async (iModelConnection: IModelConnection) => {
@@ -20,26 +17,22 @@ const ImageExportUI: FunctionComponent = () => {
     setViewportOptions({ viewState });
   };
 
-  /** The sample's render method */
   return (
     <>
-      { /** Viewport to display the iModel */}
-      {sampleIModelInfo?.iModelName && sampleIModelInfo?.contextId && sampleIModelInfo?.iModelId &&
+      { /* Viewport to display the iModel */}
+      {sampleIModelInfo?.contextId && sampleIModelInfo?.iModelId &&
         <Viewer
           contextId={sampleIModelInfo.contextId}
           iModelId={sampleIModelInfo.iModelId}
           authConfig={{ oidcClient: AuthorizationClient.oidcClient }}
           viewportOptions={viewportOptions}
-          onIModelConnected={_oniModelReady}
           defaultUiConfig={default3DSandboxUi}
           theme="dark"
-          uiProviders={uiProviders}
+          onIModelConnected={_oniModelReady}
         />
       }
     </>
   );
-
 };
 
-export default ImageExportUI;
-
+export default ViewportOnlyApp;
